jormungandr: fix after PR comments
