+++ source/jormungandr/jormungandr/realtime_schedule/tests/utils.py
@@ -1,0 +1,46 @@
+# coding=utf-8
+
+# Copyright (c) 2001-2016, Canal TP and/or its affiliates. All rights reserved.
+#
+# This file is part of Navitia,
+#     the software to build cool stuff with public transport.
+#
+# Hope you'll enjoy and contribute to this project,
+#     powered by Canal TP (www.canaltp.fr).
+# Help us simplify mobility and open public transport:
+#     a non ending quest to the responsive locomotion way of traveling!
+#
+# LICENCE: This program is free software; you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
+#
+# Stay tuned using
+# twitter @navitia
+# IRC #navitia on freenode
+# https://groups.google.com/d/forum/navitia
+# www.navitia.io
+from __future__ import absolute_import, print_function, unicode_literals, division
+
+class MockRoutePoint(object):
+    def __init__(self, *args, **kwars):
+        self._hardcoded_line_id = kwars['line_id']
+        self._hardcoded_stop_id = kwars['stop_id']
+        self._hardcoded_route_id = kwars['route_id']
+
+    def fetch_stop_id(self, rt_proxy_id):
+        return self._hardcoded_stop_id
+
+    def fetch_line_id(self, rt_proxy_id):
+        return self._hardcoded_line_id
+
+    def fetch_route_id(self, rt_proxy_id):
+        return self._hardcoded_route_id
